requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
ccxt==1.41.67
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
idna==2.10
int-date==0.1.8
multidict==5.1.0
numpy==1.20.1
pandas==1.2.2
pycares==3.1.1
pycparser==2.20
pyotp==2.4.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
stockstats==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.1.0
